+++ host/lib/usrp/cores/i2c_core_200.cpp
@@ -73,7 +73,7 @@ public:
     }
 
     void write_i2c(
-        boost::uint8_t addr,
+        boost::uint16_t addr,
         const byte_vector_t &bytes
     ){
         this->poke(REG_I2C_WR_DATA, (addr << 1) | 0); //addr and read bit (0)
@@ -96,7 +96,7 @@ public:
     }
 
     byte_vector_t read_i2c(
-        boost::uint8_t addr,
+        boost::uint16_t addr,
         size_t num_bytes
     ){
         byte_vector_t bytes;
